Switch register access to use UnsafeCell.
This seems like the least bad option at this time.  UnsafeCell is not
technically repr(C), so there is no guarantee that registers are the
right size, nor that structs built of registers describe hardware the
way we want.  Currently, it works, but I believe it's undefined.

Previously, *all* of it was undefined, because we were abusing &mut
unique references as shared.  I've gotten increasingly nervous about the
freedom that this allows the compiler; while the code it generates is
correct today, I think we were technically giving it license to break
the code later.

This feels slightly more likely to me than UnsafeCell growing a member.