stm32f4xx/rcc: change peripheral modeling.
I've gone back to the approach I used in C++, of encoding information
about enable/reset/lp bits in the peripheral enum.  This is a little
funny, involves lots of boilerplate (mostly wrapped up in a macro here),
but eliminates any need for tables in Flash and generates really tight
code.