Rework startup process.
I moved data/BSS initialization out of Rust code and into assembler,
because I'm not confident that I can predict the code's instruction set
and memory usage.  In particular, should it decide to use VFP
instructions to do the initialized data move -- not at all a bad idea!
-- the code would stop working, because the FPU isn't on yet.

To support early-but-not-that-early initialization, I've added a concept
of init hooks, like the ones used by many a crt0.  Currently I'm using
them to turn on the FPU, though technically the way I'm doing that is
also dangerous (as it's in Rust).

This merits more thought.